requirements.txt
discord.py>=2.3.0
anthropic>=0.25.0
python-dotenv>=1.0.0
